Header now added in comments so you can see article title and navigate to article

+++ app/src/main/java/com/jamie/hn/comments/domain/CommentsUseCase.kt
@@ -8,7 +8,7 @@ class CommentsUseCase(
     private val repository: StoriesRepository
 ) {
     suspend fun retrieveComments(
-        storyId: Long,
+        storyId: Int,
         useCache: Boolean,
         onResult: (List<CommentWithDepth>, Boolean, Boolean) -> Unit,
         requireComments: Boolean
